# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/ai/runs/{runId}/events:
    get:
      operationId: streamAIRunEvents
      tags: [AI Assistant]
      summary: Replay and stream durable run events
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string } }
        - { name: after, in: query, required: false, schema: { type: integer, format: int64, minimum: 0 } }
        - { name: Last-Event-ID, in: header, required: false, schema: { type: string } }
      responses:
        "200": { description: SSE event stream, content: { text/event-stream: { schema: { type: string } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna OAuth access token or personal access token
      description: Luna OAuth access token or personal access token. Individual operations may require OAuth and explicitly reject personal access tokens.
    AIAgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: independent Agent callback service credential
    AIDelegationToken:
      type: http
      scheme: bearer
      bearerFormat: 60-second operation-bound AI delegation
...
  responses:
    Unauthorized:
      description: A valid interactive browser session is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        error:
          type: string
          description: Safe user-facing error message.
        detail:
          type: string
          description: Development-only diagnostic detail.
        purpose:
          type: string
          description: Step-up purpose returned with `mfa_required`.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable error context. Scope failures include `requiredScope`.
      required: [code, error]
...